requirements.txt
PyYAML==3.11
RPi.GPIO==0.6.2
falcon==1.0.0
gunicorn==19.6.0
python-mimeparse==1.5.2
six==1.10.0
